Fix (Load  all Clusters) Add(install crone in docker + Test command)

+++ app/Console/Commands/WbMain.php
@@ -12,26 +12,31 @@ class WbMain extends Command
 
     public function handle()
     {
-        $this->info('Rolling back all migrations and reseeding the database...');
-        Artisan::call('migrate:fresh --seed');
-        $this->info('Migrations rolled back and database reseeded.');
+        if ($this->confirm('Вы уверены, что хотите запустить wb:main? Команда перезапустит все миграции вы потеряете все данные!')) 
+        {
+            $this->info('Rolling back all migrations and reseeding the database...');
+            Artisan::call('migrate:fresh --seed');
+            $this->info('Migrations rolled back and database reseeded.');
 
-        $this->info('Syncing products from Wildberries...');
-        Artisan::call('wb:get-products');
-        $this->info('Products synchronized successfully.');
+            $this->info('Syncing products from Wildberries...');
+            Artisan::call('wb:get-products');
+            $this->info('Products synchronized successfully.');
 
-        $this->info('Syncing orders from Wildberries...');
-        Artisan::call('wb:get-orders');
-        $this->info('Orders synchronized successfully.');
+            $this->info('Syncing orders from Wildberries...');
+            Artisan::call('wb:get-orders');
+            $this->info('Orders synchronized successfully.');
 
-        $this->info('Syncing stocks from Wildberries...');
-        Artisan::call('wb:get-stocks');
-        $this->info('Stocks synchronized successfully.');
+            $this->info('Syncing stocks from Wildberries...');
+            Artisan::call('wb:get-stocks');
+            $this->info('Stocks synchronized successfully.');
 
-        $this->info('Setting clusters...');
-        Artisan::call('wb:set-clusters');
-        $this->info('Clusters set successfully.');
+            $this->info('Setting clusters...');
+            Artisan::call('wb:set-clusters');
+            $this->info('Clusters set successfully.');
 
-        $this->info('All tasks completed successfully.');
+            $this->info('All tasks completed successfully.');
+        } else {
+            $this->info('Отменено.');
+        }
     }
 }
